+++ src/fr/maxlego08/menu/loader/MenuItemStackLoader.java
@@ -114,7 +114,10 @@ public class MenuItemStackLoader extends ZUtils implements Loader<MenuItemStack>
         menuItemStack.setLore(configuration.getStringList(path + "lore"));
         menuItemStack.setDisplayName(configuration.getString(path + "name", null));
         menuItemStack.setGlowing(configuration.getBoolean(path + "glow"));
-        menuItemStack.setModelID(configuration.getString(path + "modelID", "0"));
+        menuItemStack.setModelID(configuration.getString(path + "modelID",
+                configuration.getString(path + "modelId",
+                        configuration.getString(path + "customModelId",
+                                configuration.getString(path + "customModelData", "0")))));
 
         List<String> enchants = configuration.getStringList(path + "enchants");
         Map<Enchantment, Integer> enchantments = new HashMap<>();