+++ src/fr/maxlego08/menu/zcore/enums/Message.java
@@ -55,6 +55,7 @@ public enum Message {
 	DESCRIPTION_RELOAD("Allows you to reload configuration files"),
 	DESCRIPTION_VERSION("Show plugin version"),
 	DESCRIPTION_LIST("Inventory list"),
+	DESCRIPTION_TEST_DUPE("Test dupe"),
 	DESCRIPTION_DOWNLOAD("WIP"),
 	DESCRIPTION_LOGIN("WIP"),
 	DESCRIPTION_DISCONNECT("WIP"),
@@ -69,8 +70,8 @@ public enum Message {
 	DESCRIPTION_OPEN_MAIN_MENU("Open the main menu"),
 	DESCRIPTION_CREATE("Create a new config file"),
 	
-	RELOAD("§aYou have just reloaded the configuration files."),
-	RELOAD_INVENTORY("§aYou have just reloaded the inventories files."),
+	RELOAD("§aYou have just reloaded the configuration files. §8(§7%inventories% inventories§8)"),
+	RELOAD_INVENTORY("§aYou have just reloaded the inventories files. §8(§7%inventories% inventories§8)"),
 	RELOAD_INVENTORY_FILE("§aVous have just reloaded the inventory §f%name%§a."),
 	RELOAD_COMMAND("§aYou have just reloaded the commands files."),
 	RELOAD_COMMAND_FILE("§aVous have just reloaded the command §f%name%§a."),